+++ core/src/androidTest/kotlin/ca/allanwang/kau/utils/KotterknifeTest.kt
@@ -7,10 +7,12 @@ import android.support.test.runner.AndroidJUnit4
 import android.view.View
 import android.widget.FrameLayout
 import android.widget.TextView
-import org.junit.Assert.*
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
+import kotlin.test.assertEquals
+import kotlin.test.assertNotNull
+import kotlin.test.assertNull
 
 /**
  * Created by Allan Wang on 2017-07-30.