+++ mediapicker/src/main/kotlin/ca/allanwang/kau/mediapicker/MediaItem.kt
@@ -40,6 +40,7 @@ class MediaItem(val data: MediaModel)
         super.bindView(holder, payloads)
         Glide.with(holder.itemView)
                 .load(data.data)
+                .applyMediaOptions(holder.itemView.context)
                 .listener(object : RequestListener<Drawable> {
                     override fun onLoadFailed(e: GlideException?, model: Any, target: Target<Drawable>, isFirstResource: Boolean): Boolean {
                         failedToLoad = true